+++ slim/utils/schematics_ext.py
@@ -1,3 +1,4 @@
+import json
 import string
 from typing import Dict, Type
 
@@ -23,6 +24,26 @@ class BlobType(HashType):
             if len(value) % 2 == 1:
                 value = '0' + value
             return to_bin(value)
+
+
+def _json_convert(value):
+    if isinstance(value, (bytes, str)):
+        try:
+            return json.loads(value)
+        except json.JSONDecodeError:
+            raise ConversionError('Could not interpret the value as a list')
+
+
+class JSONListType(ListType):
+    def _coerce(self, value):
+        value = _json_convert(value)
+        return super()._coerce(value)
+
+
+class JSONDictType(DictType):
+    def _convert(self, value, context, safe=False):
+        value = _json_convert(value)
+        return super()._convert(value, context, safe)
 
 
 JSON_SCHEMA_TO_TYPES = {